tools/SuperResolution/requirement.txt
clip==0.2.0
lightning==1.9.5
mmedit==1.0.0rc5
mmengine==0.10.1
networkx==2.6.3
numpy==1.21.5
opencv-python==4.5.4.60
openmim==0.3.6
Pillow==9.0.1
PyYAML==6.0
tqdm==4.63.0
mmcv==2.0.0rc4